refactor: extract in custom hook
